chore(clients/go): move mock_pb into internal
