+++ clients/go/README.md
@@ -3,8 +3,8 @@
 
 ### Testing
 
-To regenerate the gateway mock `mock_pb/mock_gateway.go` run [`mockgen`](https://github.com/golang/mock#installation):
+To regenerate the gateway mock `internal/mock_pb/mock_gateway.go` run [`mockgen`](https://github.com/golang/mock#installation):
 
 ```
-mockgen github.com/zeebe-io/zeebe/clients/go/pb GatewayClient,Gateway_ActivateJobsClient > mock_pb/mock_gateway.go
+mockgen github.com/zeebe-io/zeebe/clients/go/pb GatewayClient,Gateway_ActivateJobsClient > internal/mock_pb/mock_gateway.go
 ```
