+++ clients/go/commands/setVariables_command_test.go
@@ -16,7 +16,7 @@ package commands
 
 import (
 	"github.com/golang/mock/gomock"
-	"github.com/zeebe-io/zeebe/clients/go/mock_pb"
+	"github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
 	"github.com/zeebe-io/zeebe/clients/go/pb"
 	"github.com/zeebe-io/zeebe/clients/go/utils"
 	"testing"
