+++ clients/go/commands/updateJobRetries_command_test.go
@@ -17,7 +17,7 @@ package commands
 
 import (
 	"github.com/golang/mock/gomock"
-	"github.com/zeebe-io/zeebe/clients/go/mock_pb"
+	"github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
 	"github.com/zeebe-io/zeebe/clients/go/pb"
 	"github.com/zeebe-io/zeebe/clients/go/utils"
 	"testing"
