+++ clients/go/internal/mock_pb/mock_gateway.go
