+++ clients/go/tests/license_test.go
@@ -41,7 +41,7 @@ const license = `// Copyright © 2018 Camunda Services GmbH (info@camunda.com)
 var skip = map[string]bool{
 	// These files are generated.
 	"../pb/gateway.pb.go":        true,
-	"../mock_pb/mock_gateway.go": true,
+	"../internal/mock_pb/mock_gateway.go": true,
 }
 
 func TestLicense(t *testing.T) {
